Fix: remove MyCanvas component from Results.js
This commit removes the global MyCanvas component from Room.js passed down to Results.js as a prop. Because the global MyCanvas component enables and allows for drawing and the results page should not allow drawing, only viewing. The canvas component in the results page is a placeholder for the saved png after a user finishes drawing during the game.

* Remove global MyCanvas component in Results.js that is passed down from Room.js
* Add placeholder canvas component in Results.js until we are able to replace it with a png of a finished drawing during the game

Addresses #16