feat: Category page makes a request, Room.js now has a round counter state.
This commits includes a round counter state in Room.js to help on the generation of a random seed for
the fetching on the category page. round and set round are passed as props to the category page.
Category page now has a a hook to make a request to the backend to receive 3 random catgeories.
The requests are sent with a seed which is generated from the roomId in a number format added with the
current round to change the seed every round. The seed is needed to make all players see the same content
and not each have their own random categories.

- Room.js has round state
- round is passed as props to category page
- Category makes fetch request with seed parameter
- seed parameter is calculated bu turning roomID into a number and adding the round to it
- when response is made , categories are set to their proper elements
- slight change is made on how the category array is accessed.

addresses #42