frontend: modified 'Results.js' and 'Scoreboard.sj' to display updates to scores
This commit makes it so that the 'Results' and 'Scoreboard' pages display score updates along with modifications to the tests for both pages in the file 'Room.test.js'. It also modifies the pages 'Room.js' to allow for all pages to access the list of players and guesses as well as modifying 'Lobby.js' in order to include the score data in communications between the client and server sockets.

* adds the parameters 'players', 'setPlayers', 'guesses', and 'setGuesses' to 'Room.js'
* adds a helper function 'findCorrect' (used in conjunction with 'setCorrect') to 'Results.js'
* adds the elements 'ScoreMessage' and 'BonusMessage' to 'Results.js'
* adds the parameters 'bestTrickster', 'setBestTrickster', 'bestArtist', and 'setBestArtist' to 'Scoreboard.js'
* adds the helper function 'sortUsers' (used in conjunction with 'setBestTrickster' and 'setBestArtist) to 'Scoreboard.js'
* adds the elements 'Title' and 'GameMessage' to 'Scoreboard.js'
* modifies the useEffect defined in 'Lobby.js'
* modifies the jsx defined in 'Results.js' and 'Scoreboard.js'
* modifies the tests 'loads results page' and 'loads scoreboard page' in 'Room.test.js'
* removes the parameters 'players', 'setPlayers', 'guesses', and 'setGuesses' from 'Results.js'
* removes the parameters 'data' and 'setData' from 'Scoreboard.js'

addresses #35