frontend: add function 'submitGuesses' to Drawing page and 'changeGuesses' to Voting page
This commit adds the function 'submitGuesses' to the Drawing page that takes the text entered by users and attaches it to their ids in order for votes on that guess to be scored correctly. The commit also adds the funciton 'changeGuesses' to the Voting page, which assigns the id of the user who selected it to the guess. Guesses from Drawing are displayed as yellow buttons on the Voting page that will change the 'voterIds' property of the 'guesses' variable shared amongst rooms. The commit also modifies code in the Lobby, Room, Results, and Scoreboard pages in order to keep the data associated with players and their guesses consistent between pages. Lastly, the commit modifies the tests for the Voting page in 'Room.test.js' to account for the new functions.

* adds a function 'submitGuesses' to Drawing.js
* adds a function 'changeGuesses' to Voting.js
* adds socket logic to Voting.js and Results.js
* adds a line of code to clear the value of 'guesses' between rounds in Scoreboard.js
* modifies the JSX in Drawing.js, Voting.js, Results.js
* modifies the properties shared with pages in Room.js
* modifies the visibilty of the 'players' array in 'Lobby.js' so that it is shared with other pages
* modifies tests outlined in 'Room.test.js'
* removes the 'Grid' element defined in 'Voting.js'

addresses #36